+++ tomcat/src/main/java/org/apache/catalina/connector/http/HttpCookie.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.catalina.connector.http;
 
 import java.util.Arrays;
 import java.util.HashMap;
