+++ tomcat/src/main/java/org/apache/catalina/connector/http/HttpStatus.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.catalina.connector.http;
 
 public enum HttpStatus {
     OK(200),
