+++ tomcat/src/main/java/org/apache/catalina/engine/CatalinaServletEngine.java
@@ -15,16 +15,15 @@ import java.util.Arrays;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.UUID;
 import java.util.stream.Collectors;
 
 import org.apache.commons.lang3.StringUtils;
-import org.apache.coyote.http11.HttpCookie;
-import org.apache.coyote.http11.HttpResponse;
-import org.apache.coyote.http11.HttpStatus;
+import org.apache.catalina.connector.http.HttpCookie;
+import org.apache.catalina.connector.http.HttpResponse;
+import org.apache.catalina.connector.http.HttpStatus;
 import org.apache.coyote.http11.RequestLine;
-import org.apache.coyote.http11.Session;
-import org.apache.coyote.http11.SessionManager;
+import org.apache.catalina.session.Session;
+import org.apache.catalina.session.SessionManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
